Rename IR model builder functions and variables for clarity
